Trait allocator::PageAllocator

pub trait PageAllocator: BaseAllocator {
    const PAGE_SIZE: usize;

    // Required methods
    fn alloc_pages(
        &mut self,
        num_pages: usize,
        align_pow2: usize
    ) -> AllocResult<usize>;
    fn dealloc_pages(&mut self, pos: usize, num_pages: usize);
    fn total_pages(&self) -> usize;
    fn used_pages(&self) -> usize;
    fn available_pages(&self) -> usize;
}

Page-granularity allocator.

Required Associated Constants

const PAGE_SIZE: usize

The size of a memory page.

Required Methods

fn alloc_pages( &mut self, num_pages: usize, align_pow2: usize ) -> AllocResult<usize>

Allocate contiguous memory pages with given count and alignment.

fn dealloc_pages(&mut self, pos: usize, num_pages: usize)

Deallocate contiguous memory pages with given position and count.

fn total_pages(&self) -> usize

Returns the total number of memory pages.

fn used_pages(&self) -> usize

Returns the number of allocated memory pages.

fn available_pages(&self) -> usize

Returns the number of available memory pages.

Object Safety

This trait is not object safe.